Create provisional build system.
It is a *very* unix centric system, with little to no thought put into
being cross platform. But it's what I need to handle my development
environment for the time being.

Later on, when the result is more mature I can come back and create a
build system that is useful for multiple platforms using perhaps
different tools.

For now it's a Makefile and a couple of scripts inspired by github's
"Scripts to Rule them All" idea.

script/bootstrap should be run right after you download the code. It
creates all the directories needed for the build and downloads and
builds dependencies (for now, the test dependency "atest").

script/test runs all the tests in the battery and displays a report of
how many tests were run, number of failures and errors and rate of
success.