Isolate stdint.h so we can bypass it if needed.
There's no implementation here. We merely use the stdint.h from the
development system (right now a Linux x86-64 bits with glibc and clang)
so it's no good for compatibility reasons.

The isolation is useful for when it comes to change this scenario,
though. That will make it easier to implement whatever code is needed
with a gazillion of ifdefs inside the compatibility file later.

I'm opening an issue to handle that in the future.