Create representation for Fixnums (#5)
Now we have immediate values in the form of fixnums. That complicates
our tratment of values a bit and raises a few portability issues, but
those will be dealt with in the future. For now, we have most of the
types needed for a working machine.

On with the primitive procedures now!