Simplify the interface for initializing modules
There was little reason to spread the dependency information on
basically every single file out there. We're on the path to centralize
the dependencies on the module level.

Simply put, eval and asm will depend on common and that's it. No worries
about subsystems inside those. The initializer for each library will
know which order to initialize the subsystems and there's where
dependency information will be kept on a module basis.

So far, only the work for common is done, which leaves a lot of calls in
place. After changing eval and asm things will end up simpler.