Write tests for tokenizer.
These tests use a couple of macros to make the repetitive tasks easier
to write and, I hope, easier to read.

WITH_CONTENTS sets the global variables to refer to a tokenizer holding
a character reader pointing to a buffer with the contents being the
string it receives as a parameter.

EXPECT_TOKEN verifies that calling get_next_token on the global
tokenizer produces the given token type

EXPECT_DETAILED_TOKEN verifies that calling get_next_token on the global
tokenizer produces the given token type and that subsequently calling
get_last_token_text on the same tokenizer produces the given text.

EXPECT_EOF verifies that the global tokenizer is in an EOF state.